Fixes the flicker bug (probably)
Still not sure I really understand it, but this time there is
one Svg.g element for each piece, rather than one for each
PieceGroup. Might be some problem with the DOM not recognizing
the g if there are too many Svg.use objects in each, or something.
Anyway, as far as I can tell, this seems to have fixed it.