WIP divs
Puts each piece in its own div, containing the relevant svg stuff.
The plan is then to use HTML Events to subscribe to mouse events
and move the divs around, rather than translating with svg. If all
the svg is "static" in this way, that might speed things up.

Left to do is to subscribe to HTML mouse events, and some code
cleanup.