Transitions away from svg
Svg is only used indirectly now, for creating the clip paths of
the pieces. But the pieces themselves are html divs and the
pictures are background images of the divs. It works well up to
about 4000 pieces or so.

Unfortunately, there seems to be no way to create piece outlines
for the divs without actually incorporating svg elements, which
lowers performance significantly, so this whole thing might be
a dead end anyway.

The Svg event listeners are all replaced by browser event
subscriptions, and hit detection is rewritten as well. Even the
selection boxes are drawn with divs.

The selection boxes are still slightly buggy (sometimes the
dragging don't seem to work), and there is certainly a bit of
cleanup and refactoring needed. Not sure if I can be bothered
with any of that though, as I will probably switch to webgl
first instead.