Adds option to select your own image
This requires JS ports to be able to tell the size of the new image.
This in turns requires the program to be compiled with
elm make src/Main.elm --output=main.js
And then run through index.html.

A few other changes: The images are now in Html.img objects inside
the divs, rather than css-background images directly. The reason for this
is performance: when uploading a new image, the image is turned into a
long url string which decodes back to an image. But for some reason this
is horrendously slow for the css-background, but reasonably quick for
the img tag. Everything else I tried was completely unplayable, so I'm
glad at least this worked.

TODO:
* Move the scramble/choose image buttons to a nice sub-menu
* Add a loading screen when a new image is selected (although this might
be a separate task).
* Option to select the number of pieces for the new puzzle, rather than applying
the new image directly.