Canvas and JS ports experiments
This loads the image as a texture with a sprite for each piece.
I also experiment with the canvas composite operation mode, to
see if I can get a clipping effect. Unfortunately, this doesn't
work as intended: overlapping jigsaw pieces cause undesired effects.

The new plan is to use a JS port to pre-render each jigsaw piece
as a png image, serialize that image and then render each piece
in a single div. I can use svg to draw a pretty piece, and canvas
to crop the original image and dump the final piece image to a
string. The image can also be pre-loaded by Elm, which will then
trigger the JS port event which creates the pieces.