Adds bevel effect and load text
WIP. The loading states are a little complicated, with
different different dependencies depending on whether a new
image is loaded or if it's the first time the page is loaded.
This is a first attempt, but there are still unresolved cases.
Importantly, the resetModel and scramble functions have to be
changed.

TODO. There is also a lot of cleanup to do. Code is starting
to stink quite a bit. I might want to move path scaling to JS
just so I can get rid of the extra 'getDim' step - this would
simplify things quite a bit, as the callback sequence would
then be completely linear (I think). It would also lead to
one less port, which would be nice.