Refactors Bezier and moves scaling to JS
* The Bezier type has been changed to a type alias for three Points
in a record. This simplified some of the code.
* The piece paths are now scaled in JS instead of in Elm, which
removes the need to know the image dimensions before drawing the
pieces. The trick was to not use ctx.clip but rather to use fill with
the composite operation 'destination-in' at the end.

TODO: Remove the getDim port and make the canvas draw port return
image size instead.